chore: Reduce folder nesting by moving the config folder up
