+++ packages/font-config-webpack-plugin/config/development.config.js
@@ -1,20 +1,19 @@
 /**
  * Common Development Config
  *
- * @param {import("../ImageConfigWebpackPlugin.js").ImageConfigWebpackPluginOptions} options
+ * @param {import("../src/FontConfigWebpackPlugin.js").FontConfigWebpackPluginOptions} options
  * @returns {{ module: { rules : Array<any> }, plugins: Array<(new (): any)> }}
  */
 exports = module.exports = options => ({
 	module: {
 		rules: [
 			{
-				test: /\.(gif|jpg|jpeg|png|svg)$/,
+				test: /\.(woff2?)(\?v=\d+\.\d+\.\d+)?$/,
 				use: [
 					{
-						loader: require.resolve('url-loader'),
+						loader: require.resolve('file-loader'),
 						options: {
 							name: options.name,
-							limit: 512,
 						},
 					},
 				],
