+++ packages/font-config-webpack-plugin/config/production.config.js
@@ -1,7 +1,7 @@
 /**
  * Common Production Config
  *
- * @param {import("../FontConfigWebpackPlugin.js").FontConfigWebpackPluginOptions} options
+ * @param {import("../src/FontConfigWebpackPlugin.js").FontConfigWebpackPluginOptions} options
  * @returns {{ module: { rules : Array<any> }, plugins: Array<(new (): any)> }}
  */
 exports = module.exports = options => require('./development.config')(options);
