+++ packages/image-config-webpack-plugin/config/development.config.js
@@ -1,14 +1,14 @@
 /**
  * Common Development Config
  *
- * @param {import("../FontConfigWebpackPlugin.js").FontConfigWebpackPluginOptions} options
+ * @param {import("../src/ImageConfigWebpackPlugin.js").ImageConfigWebpackPluginOptions} options
  * @returns {{ module: { rules : Array<any> }, plugins: Array<(new (): any)> }}
  */
 exports = module.exports = options => ({
 	module: {
 		rules: [
 			{
-				test: /\.(woff2?)(\?v=\d+\.\d+\.\d+)?$/,
+				test: /\.(gif|jpg|jpeg|png|svg)$/,
 				use: [
 					{
 						loader: require.resolve('file-loader'),
