+++ packages/image-config-webpack-plugin/config/production.config.js
@@ -1,7 +1,7 @@
 /**
  * Common Development Config
  *
- * @param {import("../ImageConfigWebpackPlugin.js").ImageConfigWebpackPluginOptions} options
+ * @param {import("../src/ImageConfigWebpackPlugin.js").ImageConfigWebpackPluginOptions} options
  * @returns {{ module: { rules : Array<any> }, plugins: Array<(new (): any)> }}
  */
 exports = module.exports = options => ({
@@ -11,9 +11,10 @@ exports = module.exports = options => ({
 				test: /\.(gif|jpg|jpeg|png|svg)$/,
 				use: [
 					{
-						loader: require.resolve('file-loader'),
+						loader: require.resolve('url-loader'),
 						options: {
 							name: options.name,
+							limit: 512,
 						},
 					},
 				],
