+++ packages/scss-config-webpack-plugin/config/production.config.js
@@ -7,7 +7,7 @@ const safeParser = require('postcss-safe-parser');
  *
  * Common Production Config
  *
- * @param {import("../ScssConfigWebpackPlugin.js").ScssConfigWebpackPluginOptions} options
+ * @param {import("../src/ScssConfigWebpackPlugin.js").ScssConfigWebpackPluginOptions} options
  * @returns {{ module: { rules : Array<any> }, plugins: Array<(new (): any)> }}
  */
 exports = module.exports = options => ({
