+++ packages/scss-config-webpack-plugin/src/ScssConfigWebpackPlugin.js
@@ -29,8 +29,8 @@ class ScssConfigWebpackPlugin {
 				? this.options.mode === 'production'
 				: compiler.options.mode === 'production' || !compiler.options.mode;
 		const config = isProductionLikeMode
-			? require('./config/production.config')(this.options)
-			: require('./config/development.config')(this.options);
+			? require('../config/production.config')(this.options)
+			: require('../config/development.config')(this.options);
 		// Merge config
 		config.plugins.forEach(plugin => plugin.apply(compiler));
 		compiler.hooks.afterEnvironment.tap('ScssConfigWebpackPlugin', () => {
