+++ packages/ts-config-webpack-plugin/config/development.config.js
@@ -5,7 +5,7 @@ const ForkTsCheckerWebpackPlugin = require('fork-ts-checker-webpack-plugin');
 /**
  * Common Development Config
  *
- * @param {import("../TsConfigWebpackPlugin.js").TsConfigWebpackPluginOptions} options
+ * @param {import("../src/TsConfigWebpackPlugin.js").TsConfigWebpackPluginOptions} options
  * @returns {{ module: { rules : Array<any> }, plugins: Array<(new (): any)> }}
  */
 exports = module.exports = options => ({
