+++ packages/ts-config-webpack-plugin/config/tsconfig.base.json
